backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
PyMuPDF==1.22.5
langchain-core>=0.1.0,<0.2.0
langchain>=0.1.0
langchain-groq>=0.0.1
Werkzeug==3.0.1 